requirements.txt
mysql-connector-python==9.2.0
Pillow==11.1.0
psycopg==3.2.4
psycopg-binary==3.2.4
pyTelegramBotAPI==4.26.0
python-dotenv==1.0.1
